Implement AlbumController

+++ Business/Services/AlbumService.cs
@@ -18,17 +18,17 @@ namespace Business.Services
             _mapper = mapper;
             _musicBrainzQueryService = musicBrainzQueryService;
         }
-        public async Task AddAsync(AlbumDto model)
-        {
-            var albumInDb = await _unitOfWork.AlbumRepository.GetByIdAsync(model.Id);
-            if (albumInDb != null)
-            {
-                throw new MusicLibraryException("Album already exists");
-            }
-            var album = _mapper.Map<Album>(model);
-            await _unitOfWork.AlbumRepository.AddAsync(album);
-            await _unitOfWork.SaveChangesAsync();
-        }
+        //public async Task AddAsync(AlbumDto model)
+        //{
+        //    var albumInDb = await _unitOfWork.AlbumRepository.GetByIdAsync(model.Id);
+        //    if (albumInDb != null)
+        //    {
+        //        throw new MusicLibraryException("Album already exists");
+        //    }
+        //    var album = _mapper.Map<Album>(model);
+        //    await _unitOfWork.AlbumRepository.AddAsync(album);
+        //    await _unitOfWork.SaveChangesAsync();
+        //}
 
         public async Task AddByMusicBrainzIdAsync(Guid id)
         {
@@ -52,6 +52,11 @@ namespace Business.Services
 
         public async Task DeleteAsync(Guid modelId)
         {
+            var albumInDb = await _unitOfWork.AlbumRepository.GetByIdAsync(modelId);
+            if (albumInDb == null)
+                throw new MusicLibraryException("Album is not present in the database");
+
+            await _unitOfWork.AlbumRepository.DeleteConnectionsByAlbumIdAsync(modelId);
             await _unitOfWork.AlbumRepository.DeleteByIdAsync(modelId);
             await _unitOfWork.SaveChangesAsync();
         }
